+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Add: Format Dates, Date Math, Extract Text, Clean Text and Fill Empty Values operations.
+- Add: Format Dates, Date Math, Extract Text, Clean Text and Fill Empty Values operations (5a32941).
 
 ## 1.2.0